backend/requirements.txt
fastapi
uvicorn
python-dotenv
openai
anthropic
supabase
python-multipart
pydantic
GitPython
pyyaml
huggingface_hub
requests
groq
gunicorn
